Collegio apps using branch moka coffee. Is the collegio apps that using eikon black nodes. Add apps for running multiple apps on node. Add app java for burn node id. Fix bug in node id eeprom read. Update ADC libraries, Light libraries.

+++ VirtualSense/DJ_VirtualMachine.1.0/darjeeling.1.1/src/lib/virtualsense/javax/virtualsense/sensors/Light.java
@@ -29,10 +29,9 @@ package javax.virtualsense.sensors ;
 public class Light
 { 
 	/**
-	 * Reads light value from the BH1620FVC sensor.	 
-	 * @return measured light express in mLx.
-	 */
-	//TODO: c'è un motivo per ritornare un int e non uno short ?
-	public static native int getValue();
+     * Reads light value from the BH1620FVC sensor.  
+     * @return measured light express in Lx (Lux).
+     */
+    public static native short getValue();
 
 }
